fix: Use Long constants in Minimum/Maximum clear() methods
The Minimum and Maximum aggregation functions had a type mismatch bug
in their clear() methods:
- Minimum.clear() used Integer.MAX_VALUE instead of Long.MAX_VALUE
- Maximum.clear() used Integer.MIN_VALUE instead of Long.MIN_VALUE

This caused incorrect behavior when tracking values outside the 32-bit
integer range after a clear() operation.

Impact:
- After clear(), comparisons with values > Integer.MAX_VALUE would fail
  in Minimum because Integer.MAX_VALUE < large values
- After clear(), comparisons with values < Integer.MIN_VALUE would fail
  in Maximum because Integer.MIN_VALUE > very negative values

This fix ensures that clear() properly resets to the correct sentinel
values for 64-bit long arithmetic.

Added comprehensive test suites for both classes covering:
- Initial values and simple operations
- Values exceeding Integer range
- Clear and reuse scenarios
- Boundary values (Long.MIN_VALUE, Long.MAX_VALUE)
- Edge cases (zero, negative, mixed values)

All 26 tests pass.

+++ btrace-core/src/main/java/org/openjdk/btrace/core/aggregation/Minimum.java
@@ -37,7 +37,7 @@ class Minimum implements AggregationValue {
 
   @Override
   public synchronized void clear() {
-    min = Integer.MAX_VALUE;
+    min = Long.MAX_VALUE;
   }
 
   @Override
